kotlin nmsl 2.0

+++ src/minecraft/kevin/module/modules/world/OldScaffold.kt
@@ -1271,12 +1271,16 @@ class OldScaffold : ClientModule("OldScaffold", "scaffolding", ModuleCategory.WO
     private fun calculateRotation(rotation: Rotation): Rotation {
         return if (towerState) when(towerRotationsValue.get()) {
             "Custom" -> Rotation(mc.thePlayer.rotationYaw + customTowerYawValue.get(), customTowerPitchValue.get())
-            "MoveDirection" -> Rotation(MovementUtils.movingYaw - 180, rotation.pitch)
+            "MoveDirection" -> Rotation(MovementUtils.movingYaw - 180,
+                rotation.pitch.toDouble()
+            )
             else -> rotation
         } else when(rotationsValue.get()) {
             "AAC" -> Rotation(mc.thePlayer.rotationYaw + (((if (mc.thePlayer.movementInput.moveForward < 0) 0 else 180) - aacYawOffsetValue.get()) * if (aacRotationPositive) 1 else -1), rotation.pitch)
             "LimitedAAC" -> Rotation(mc.thePlayer.rotationYaw + (((if (mc.thePlayer.movementInput.moveForward < 0) 0 else 180) - aacYawOffsetValue.get()) * if (aacRotationPositive) 1 else -1), customPitchValue.get())
-            "MoveDirection" -> Rotation(MovementUtils.movingYaw - 180, customPitchValue.get())
+            "MoveDirection" -> Rotation(MovementUtils.movingYaw - 180,
+                customPitchValue.get().toDouble()
+            )
             "Custom" -> Rotation(mc.thePlayer.rotationYaw + customYawValue.get(), customPitchValue.get())
             else -> rotation
         }
